replace pseudo HOC with render prop
